Refactored the roc_curve tests out into a separate test_roc_curve.py
